[API] correct log and extract log from job dependencies builder
